templates/gateway/requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
Django==5.0.7
kafka-python==2.0.2
kafka-python-ng==2.2.2
msgpack==1.0.8
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.43.0
Flask==2.3.2
requests==2.31.0
gunicorn==21.2.0
